+++ NEWS.md
@@ -5,6 +5,7 @@
 * `merge_v` has a new argument `combine` to let use j columns 
 be used as a single value (all values are pasted).
 * new function `add_body` for adding rows into a flextable body
+* new function `colformat_image` for images in flextable
 * new method `as_flextable` for `gam` models
 
 ## Issues